Exercise 42 - Custom Hooks - 3
