+++ React/src/Components/GitHubUser/GitHubUser.jsx
@@ -1,22 +1,9 @@
-import { useEffect, useState } from "react";
+import useGithubUser from "../CustomHooks/useGitHubUser";
 
 export function GitHubUser({ username }) {
 
-  const [data, setData] = useState(null);
-
-  async function fetchGitHubUser(username) {
-    try {
-      const response = await fetch(`https://api.github.com/users/${username}`);
-      const json = await response.json();
-      console.log(json);
-      setData(json);
-    } catch (error) {
-      console.log("Error", error);
-    }
-  }
-  useEffect(() => {
-    fetchGitHubUser(username);
-  }, [username]);
+  const data = useGithubUser(username);
+  
   return (
     <div>
       {data && (
